components split up in adminhub
