+++ src/components/AdminHub/AdminHub.jsx
@@ -1,30 +1,19 @@
 import { useEffect, useState } from "react";
 import { useHistory } from "react-router";
 import { useSelector, useDispatch } from "react-redux";
-import {
-  Grid,
-  Typography,
-  Container,
-  Button,
-  Paper,
-  TextField,
-  IconButton,
-} from "@mui/material";
-import Table from "@mui/material/Table";
-import TableBody from "@mui/material/TableBody";
-import TableCell from "@mui/material/TableCell";
-import TableContainer from "@mui/material/TableContainer";
-import TableHead from "@mui/material/TableHead";
-import TableRow from "@mui/material/TableRow";
-
-import ToggleOffIcon from "@mui/icons-material/ToggleOff";
+import { Grid, Typography, Container, Button, TextField } from "@mui/material";
 
 import NewPostingsReady from "../NewJobOpeningsReadyAlert/NewJobOpeningsReadyAlert";
 import NewSubmissions from "../NewSubmissionsAlert/NewSubmissionsAlert";
 import NewIssues from "../NewIssusAlert/NewIssuesAlert";
-import useStyles from "../Styles/Styles"; //important paste this
+import useStyles from "../Styles/Styles";
+
+import AdminHubFourColTable from "./AdminHubFourColTable";
+import AdminHubThreeColTable from "./AdminHubThreeColTable";
+import AdminHubTwoColTable from "./AdminHubTwoColTable";
 
 import "./AdminHub.css";
+import AdminHubManualSubEntryForm from "./AdminHubManualSubEntryForm";
 
 const AdminHub = () => {
   const history = useHistory();
@@ -37,12 +26,6 @@ const AdminHub = () => {
   const [isLoading, setIsLoading] = useState(false);
   const [subList, setSubList] = useState([]);
 
-  //manual user entry info
-  const [userEmail, setUserEmail] = useState("");
-  const [userFirstName, setUserFirstName] = useState("");
-  const [userLastName, setUserLastName] = useState("");
-  const [userZip, setUserZip] = useState("");
-
   //user sub status
   const [subStatus, setSubStatus] = useState(""); //deletelater
 
@@ -53,45 +36,6 @@ const AdminHub = () => {
     setSubList(subs.data);
     console.log("user is:", user.first_name);
   }, [subList]);
-
-  //check email - is it real? if keep required attribute
-  const validateEmail = () => {
-    if (userEmail.indexOf("@") > -1) {
-      submit();
-    } else {
-      alert("invalid email!");
-    }
-  };
-
-  // new subscriber admin post!
-  const submit = () => {
-    const userAddress = {
-      addr1: "test addr",
-      city: "test city",
-      state: "test state",
-      zip: userZip,
-    };
-    console.log(
-      "valid!! sending:",
-      userEmail,
-      userFirstName,
-      userLastName,
-      userZip
-    );
-    dispatch({
-      type: "ADD_SUBSCRIBER",
-      payload: {
-        email: userEmail,
-        firstName: userFirstName,
-        lastName: userLastName,
-        address: userAddress,
-      },
-    });
-    setUserEmail("");
-    setUserFirstName("");
-    setUserLastName("");
-    setUserZip("");
-  };
 
   const toggleSubStatus = (subStatus, id) => {
     let newStatus = "";
@@ -111,10 +55,7 @@ const AdminHub = () => {
 
   return (
     <div className="adminHubPage">
-      <Typography
-
-        className={classes.adminHeader}
-      >
+      <Typography className={classes.adminHeader}>
         Hi {user.first_name}!
       </Typography>
 
@@ -153,309 +94,14 @@ const AdminHub = () => {
               </div>
             </Grid>
           </Grid>
-          <div>
-            <Typography variant="h4" className={classes.subEntryHeader}>Manual Subscriber Entry</Typography>
-            <form onSubmit={() => validateEmail()}>
-              <TextField
-                value={userEmail}
-                type="email"
-                id="email"
-                size="small"
-                placeholder="Email"
-                style={{ width: "200px", margin: '4px' }}
-                onChange={(event) => setUserEmail(event.target.value)}
-              />{" "}
-              <br />
-              <TextField
-                value={userFirstName}
-                type="text"
-                id="firstName"
-                size="small"
-                placeholder="First Name"
-                style={{ width: "200px", margin: '4px'  }}
-                onChange={(event) => setUserFirstName(event.target.value)}
-              />
-              <br />
-              <TextField
-                value={userLastName}
-                type="text"
-                id="lastName"
-                size="small"
-                placeholder="Last Name"
-                style={{ width: "200px", margin: '4px'  }}
-                onChange={(event) => setUserLastName(event.target.value)}
-              />
-              <br />
-              <TextField
-                value={userZip}
-                type="text"
-                id="email"
-                size="small"
-                placeholder="Zip Code"
-                style={{ width: "200px", margin: '4px'  }}
-                onChange={(event) => setUserZip(event.target.value)}
-              />
-              <br />
-              <Button
-                style={{ margin: "15px" }}
-    
-                variant="contained"
-                type="submit"
-              >
-                Submit
-              </Button>
-            </form>
-            <br />
-          </div>
 
-          {/* ---------- FOUR COLUMN TABLE ON 3000 PX (AKA nothing is too big, hopefully) ---------- */}
-
-
-          <div className="fourColTable">
-            <Typography
-              variant="h4"
-              className={classes.fullSubListHeader}
-            >
-              Full Subscriber List:
-            </Typography>
-
-            <Paper elevation={4} className="adminPaper" >
-              <Grid>
-                <TableContainer sx={{ maxHeight: 470 }}>
-                  <Table
-                    className="tableMain"
-                    stickyHeader
-                    aria-label="sticky table"
-                  >
-                    <TableHead className={classes.tableHeader}>
-                      <TableRow>
-                        <TableCell className={classes.tableHeaderCell}>
-                          Name
-                        </TableCell>
-                        <TableCell className={classes.tableHeaderCell}>
-                          Email
-                        </TableCell>
-                        <TableCell className={classes.tableHeaderCell}>
-                          Zip Code
-                        </TableCell>
-                        <TableCell className={classes.tableHeaderCell}>
-                          Status
-                        </TableCell>
-                      </TableRow>
-                    </TableHead>
-                    <TableBody className="adminTableBody">
-                      {subs.length > 0 ? (
-                        subs[0].map((item) => (
-                          <TableRow
-                            key={item.id}
-                            className={classes.tableBodyRow}
-                          >
-                            <TableCell className={classes.tableCell}>
-                              {item.full_name}
-                            </TableCell>
-                            <TableCell className={classes.tableCell}>
-                              {item.email_address}
-                            </TableCell>
-
-                            <TableCell className={classes.tableCell}>
-                              {item.merge_fields.ADDRESS.zip}
-                            </TableCell>
-
-                            <TableCell className={classes.tableCell}>
-                              {item.status}
-
-                              <IconButton
-                              onClick={() =>
-                                toggleSubStatus(item.status, item.contact_id)
-                              }
-                            >
-                              <ToggleOffIcon />
-                            </IconButton>
-
-
-                            </TableCell>
-                            
-                          </TableRow>
-                        ))
-                      ) : (
-                        <img src="./images/Pendulum.gif" />
-                      )}
-                    </TableBody>
-                  </Table>
-                </TableContainer>
-              </Grid>
-            </Paper>
-        
-          </div>
-
-{/* ---------- THREE COLUMN TABLE ON 600 PX ---------- */}
-
-
-<div className="threeColTable">
-            <Typography
-              variant="h4"
-              className={classes.fullSubListHeader}
-            >
-              Full Subscriber List:
-            </Typography>
-
-            <Paper elevation={4} className="adminPaper" >
-              <Grid>
-                <TableContainer sx={{ maxHeight: 470 }}>
-                  <Table
-                    className="tableMain"
-                    stickyHeader
-                    aria-label="sticky table"
-                  >
-                    <TableHead className={classes.tableHeader}>
-                      <TableRow>
-                        <TableCell className={classes.tableHeaderCell}>
-                          Name
-                        </TableCell>
-                        <TableCell className={classes.tableHeaderCell}>
-                          Email
-                        </TableCell>
-                        {/* <TableCell className={classes.tableHeaderCell}>
-                          Zip Code
-                        </TableCell> */}
-                        <TableCell className={classes.tableHeaderCell}>
-                          Status
-                        </TableCell>
-                      </TableRow>
-                    </TableHead>
-                    <TableBody className="adminTableBody">
-                      {subs.length > 0 ? (
-                        subs[0].map((item) => (
-                          <TableRow
-                            key={item.id}
-                            className={classes.tableBodyRow}
-                          >
-                            <TableCell className={classes.tableCell}>
-                              {item.full_name}
-                            </TableCell>
-                            <TableCell className={classes.tableCell}>
-                              {item.email_address}
-                            </TableCell>
-
-                            {/* <TableCell className={classes.tableCell}>
-                              {item.merge_fields.ADDRESS.zip}
-                            </TableCell> */}
-
-                            <TableCell className={classes.tableCell}>
-                              {item.status}
-
-                              <IconButton
-                              onClick={() =>
-                                toggleSubStatus(item.status, item.contact_id)
-                              }
-                            >
-                              <ToggleOffIcon />
-                            </IconButton>
-
-
-                            </TableCell>
-                            
-                          </TableRow>
-                        ))
-                      ) : (
-                        <img src="./images/Pendulum.gif" />
-                      )}
-                    </TableBody>
-                  </Table>
-                </TableContainer>
-              </Grid>
-            </Paper>
-        
-          </div>
-
-{/* ---------- TWO COLUMN TABLE ON 600 PX ---------- */}
-
-
-<div className="twoColTable">
-            <Typography
-              variant="h4"
-              className={classes.fullSubListHeader}
-            >
-              Full Subscriber List:
-            </Typography>
-
-            <Paper elevation={4} className="adminPaper" >
-              <Grid>
-                <TableContainer sx={{ maxHeight: 470 }}>
-                  <Table
-                    className="tableMain"
-                    stickyHeader
-                    aria-label="sticky table"
-                  >
-                    <TableHead className={classes.tableHeader}>
-                      <TableRow>
-                        {/* <TableCell className={classes.tableHeaderCell}>
-                          Name
-                        </TableCell> */}
-                        <TableCell className={classes.tableHeaderCell}>
-                          Email
-                        </TableCell>
-                        {/* <TableCell className={classes.tableHeaderCell}>
-                          Zip Code
-                        </TableCell> */}
-                        <TableCell className={classes.tableHeaderCell}>
-                          Status
-                        </TableCell>
-                      </TableRow>
-                    </TableHead>
-                    <TableBody className="adminTableBody">
-                      {subs.length > 0 ? (
-                        subs[0].map((item) => (
-                          <TableRow
-                            key={item.id}
-                            className={classes.tableBodyRow}
-                          >
-                            {/* <TableCell className={classes.tableCell}>
-                              {item.full_name}
-                            </TableCell> */}
-                            <TableCell className={classes.tableCell}>
-                              {item.email_address}
-                            </TableCell>
-
-                            {/* <TableCell className={classes.tableCell}>
-                              {item.merge_fields.ADDRESS.zip}
-                            </TableCell> */}
-
-                            <TableCell className={classes.tableCell}>
-                              {item.status}
-
-                              <IconButton
-                              onClick={() =>
-                                toggleSubStatus(item.status, item.contact_id)
-                              }
-                            >
-                              <ToggleOffIcon />
-                            </IconButton>
-
-
-                            </TableCell>
-                            
-                          </TableRow>
-                        ))
-                      ) : (
-                        <img src="./images/Pendulum.gif" />
-                      )}
-                    </TableBody>
-                  </Table>
-                </TableContainer>
-              </Grid>
-            </Paper>
-        
-          </div>
-
-
-
-
-
-
-
-
-
+          <AdminHubManualSubEntryForm />
+          <AdminHubFourColTable subs={subs} toggleSubStatus={toggleSubStatus} />
+          <AdminHubThreeColTable
+            subs={subs}
+            toggleSubStatus={toggleSubStatus}
+          />
+          <AdminHubTwoColTable subs={subs} toggleSubStatus={toggleSubStatus} />
         </div>
       </Container>
     </div>
